# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "The Assessment Tracker is an application designed to help track an individual's progress through a batch training. A user can see all the batches organized by year, view a particular week for a batch, create new assessments, and organize assessments into types. Each assessment is assigned a weight to contribute to the final grade, and associates that complete a given assignment are given a grade in the application.",
    "version": "1.0.0",
    "title": "Assessment-Tracker-Python-Endpoints",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
...
  "paths": {
...
    "/login": {
      "post": {
        "tags": ["instructor"],
        "summary": "Login an instructor",
        "requestBody": {
          "description": "Login an instructor using their email",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginTrainer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trainer successfully logs in"
          },
          "404": {
            "description": "Trainer unsuccessfully logs in"
          }
        }
      }
    },
    "/trainers": {
      "post": {
        "tags": ["admin"],
        "summary": "Create a new trainer.",
        "requestBody": {
          "description": "Create a new trainer.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              },
              "examples": {
                "1": {
                  "value": {
                    "firstName": "Trainer",
                    "lastName": "McTrainerface",
                    "email": "important@trainer.person"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/trainer"
                },
                "examples": {
                  "recieve trainer": {
                    "value": {
                      "email": "important@trainer.person",
                      "firstName": "Trainer",
                      "id": 1,
                      "lastName": "McTrainerface",
                      "role": "",
                      "admin": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/trainers/register": {
      "post": {
        "tags": ["admin"],
        "summary": "Assign trainer to batch.",
        "requestBody": {
          "description": "Create a new trainer-batch join relationship.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "trainerId": {
                    "type": "integer"
                  },
                  "batchId": {
                    "type": "integer"
                  },
                  "trainerRole": {
                    "type": "string"
                  }
                }
              },
              "examples": {
                "1": {
                  "value": {
                    "trainerId": 1,
                    "batchId": 1,
                    "trainerRole": "Lead"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/trainer"
                },
                "examples": {
                  "recieve trainer": {
                    "value": {
                      "email": "important@trainer.person",
                      "firstName": "Trainer",
                      "id": 1,
                      "lastName": "McTrainerface",
                      "role": "Lead",
                      "admin": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/trainers/{trainer_id}": {
      "parameters": [
        {
          "name": "trainer_id",
          "in": "path",
          "required": true,
          "description": "Trainer Id",
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": ["admin"],
        "summary": "Takes in an id and returns a Trainer",
        "responses": {
          "200": {
            "description": "Trainer found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getTrainerResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Trainer not found"
          },
          "400": {
            "description": "Bad request!"
          }
        }
      }
    },
    "/batches/{batch_id}/trainers": {
      "parameters": [
        {
          "name": "batch_id",
          "in": "path",
          "required": true,
          "description": "Batch Id",
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": ["admin"],
        "summary": "Takes in an id for a batch record and returns a Trainer or trainers",
        "responses": {
          "200": {
            "description": "Successfully found Trainer(s)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/getTrainerResponseBody"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/years?trainerId=1": {
      "parameters": [
        {
          "name": "trainerId",
          "in": "path",
          "required": true,
          "description": "Trainer Id",
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": ["admin"],
        "summary": "returns the years a trainer was active",
        "responses": {
          "200": {
            "description": "Returns the years a trainer was active based on the Id provided",
            "content": {
              "application/json": {
                "type": "array",
                "items": {
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "description": "Bad request!"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "getTrainerResponseBody": {
        "properties": {
          "id": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "admin": {
            "type": "boolean"
          }
        }
      },
...
      "loginTrainer": {
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
...
      "trainer": {
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "email": {
            "type": "string"
          },
          "admin": {
            "type": "boolean"
          }
        }
      }
    }
  }
}